Refactoring code : STudent D
