+++ classes/item.rb
@@ -1,8 +1,10 @@
 require 'date'
 
 class Item
+  attr_accessor :archived, :source, :publish_date, :id
+
   def initialize(publish_date, id = nil)
-    @id = id || random.rand(1..1000)
+    @id = id || Random.rand(1..1000)
     @publish_date = publish_date
   end
 
@@ -12,7 +14,10 @@ class Item
 
   def add_label; end
 
-  def add_source; end
+  def add_source(source)
+    @source = source
+    source.items << self unless source.items.include?(self)
+  end
 
   def can_be_archived?
     current_year = Date.today.year
