+++ classes/movie/movie.rb
@@ -1,7 +1,7 @@
-require_relative '../item_movie'
+require_relative '../item'
 require_relative 'source'
 
-class Movie < ItemMovie
+class Movie < Item
   attr_accessor :archived, :silent
 
   def initialize(publish_date, silent:, id: nil)
