+++ data/movie.json
@@ -10,5 +10,11 @@
     "publish_date": "2003/02/03",
     "silent": false,
     "source": "Online Shop"
+  },
+  {
+    "id": 376,
+    "publish_date": "2003/05/04",
+    "silent": true,
+    "source": "Netflix"
   }
 ]
