+++ main.rb
@@ -1,40 +1,41 @@
-def list_options
-  list = "
-1- List books
-2- List games
-3- List movies
-4- List music
-5- Exit
-"
-  puts list
-  gets.chomp
-end
+require_relative 'app'
 
-def switch(number)
-  loop do
-    case number
-    when '1'
-      puts 'option 1'
-      number = list_options
-    when '2'
-      puts 'option 2'
-      number = list_options
-    when '3'
-      puts 'option 3'
-      number = list_options
-    when '4'
-      puts 'option 4'
-      number = list_options
-    else
-      break
+class Main
+  def initialize
+    @app = App.new
+    prompt
+  end
+
+  def prompt
+    puts 'Catalogue of things!'
+    loop do
+      display_options
+      option = take_input
+      @app.run(option)
+
+      break if option == 13
     end
+  end
+
+  def display_options
+    puts 'Please choose one of the options: '
+
+    puts '3. - List all movies'
+    puts '8. - List all sources'
+    puts '11. - Add movie'
+    puts '13. - Exit'
+  end
+
+  def take_input
+    print 'Enter a number: '
+    option = gets.chomp.to_i
+    while option.nil? || option < 1 || option > 13
+      puts 'Please enter a number between 1 and 4!'
+      print 'Enter a number: '
+      option = gets.chomp.to_i
+    end
+    option
   end
 end
 
-def main
-  puts 'Welcome to School Library App!'
-  number = list_options
-  switch(number)
-end
-
-main
+Main.new
