+++ spec/book_spec.rb
