+++ spec/label_spec.rb
