+++ spec/movie_spec.rb
