+++ spec/music_spec.rb
